+++ lua/obsidian/commands/init-legacy.lua
@@ -24,7 +24,6 @@ local command_lookups = {
   ObsidianRename = "obsidian.commands.rename",
   ObsidianPasteImg = "obsidian.commands.paste_img",
   ObsidianExtractNote = "obsidian.commands.extract_note",
-  ObsidianDebug = "obsidian.commands.debug",
   ObsidianTOC = "obsidian.commands.toc",
 }
 
@@ -180,8 +179,6 @@ M.register(
   "ObsidianExtractNote",
   { opts = { nargs = "?", range = true, desc = "Extract selected text to a new note and link to it" } }
 )
-
-M.register("ObsidianDebug", { opts = { nargs = 0, desc = "Log some information for debugging" } })
 
 M.register("ObsidianTOC", { opts = { nargs = 0, desc = "Load the table of contents into a picker" } })